Added Snack handling and synchronization with iZettle after model updates
This commit adds handling of Snack records, such as creating, listing, and editing. It also adds synchronization with iZettle whenever a company's model is updated. The commit also modifies several files to ensure that the appropriate access control is in place for various routes. For instance, company admin middleware is applied to certain create and edit routes.